switch to color-lib and color-validation submodules
