+++ apps/bundle-size/baseline.json
@@ -69,32 +69,32 @@
     },
     {
       "platform": "macos",
-      "scenario": "design-runtime-colors",
-      "rawBytes": 825935,
-      "gzipBytes": 238683,
-      "moduleCount": 499,
-      "metroModuleCount": 497,
-      "metafileInputCount": 499,
+      "scenario": "design-color-lib",
+      "rawBytes": 826875,
+      "gzipBytes": 239247,
+      "moduleCount": 503,
+      "metroModuleCount": 501,
+      "metafileInputCount": 503,
       "workspaceModules": {
-        "@fluentui-react-native/design": 2
+        "@fluentui-react-native/design": 6
       },
       "workspaceBytes": {
-        "@fluentui-react-native/design": 5327
+        "@fluentui-react-native/design": 6213
       }
     },
     {
       "platform": "macos",
-      "scenario": "design-contrast",
-      "rawBytes": 827111,
-      "gzipBytes": 238439,
-      "moduleCount": 499,
-      "metroModuleCount": 497,
-      "metafileInputCount": 499,
+      "scenario": "design-color-validation",
+      "rawBytes": 827655,
+      "gzipBytes": 238605,
+      "moduleCount": 505,
+      "metroModuleCount": 503,
+      "metafileInputCount": 507,
       "workspaceModules": {
-        "@fluentui-react-native/design": 2
+        "@fluentui-react-native/design": 8
       },
       "workspaceBytes": {
-        "@fluentui-react-native/design": 6502
+        "@fluentui-react-native/design": 7012
       }
     },
     {
@@ -199,32 +199,32 @@
     },
     {
       "platform": "windows",
-      "scenario": "design-runtime-colors",
-      "rawBytes": 835070,
-      "gzipBytes": 240083,
-      "moduleCount": 504,
-      "metroModuleCount": 502,
-      "metafileInputCount": 504,
+      "scenario": "design-color-lib",
+      "rawBytes": 836010,
+      "gzipBytes": 240687,
+      "moduleCount": 508,
+      "metroModuleCount": 506,
+      "metafileInputCount": 508,
       "workspaceModules": {
-        "@fluentui-react-native/design": 2
+        "@fluentui-react-native/design": 6
       },
       "workspaceBytes": {
-        "@fluentui-react-native/design": 5327
+        "@fluentui-react-native/design": 6213
       }
     },
     {
       "platform": "windows",
-      "scenario": "design-contrast",
-      "rawBytes": 836244,
-      "gzipBytes": 239875,
-      "moduleCount": 504,
-      "metroModuleCount": 502,
-      "metafileInputCount": 504,
+      "scenario": "design-color-validation",
+      "rawBytes": 836790,
+      "gzipBytes": 240087,
+      "moduleCount": 510,
+      "metroModuleCount": 508,
+      "metafileInputCount": 512,
       "workspaceModules": {
-        "@fluentui-react-native/design": 2
+        "@fluentui-react-native/design": 8
       },
       "workspaceBytes": {
-        "@fluentui-react-native/design": 6502
+        "@fluentui-react-native/design": 7012
       }
     },
     {
