+++ packages/agentic/design/src/color-lib/index.ts
@@ -1,0 +1,33 @@
+export type {
+  ColorDiagnostic,
+  ColorDiagnosticReason,
+  ColorMode,
+  InteractionColorDiagnostic,
+  InteractionColorInput,
+  InteractionColorOptions,
+  InteractionColorOverridesResult,
+  InteractionColorResult,
+  InteractionColorsResult,
+  InteractionColorState,
+  InteractionColorVariant,
+  InteractiveColorToken,
+  OklchColor,
+  ParsedColorValue,
+  ResolvedContrastColors,
+  RgbColor,
+  RgbaColor,
+} from './types';
+export { parseColorValue, rgbaToHex } from './parsing';
+export { oklchToRgb, rgbToOklch } from './conversion';
+export { compositeColor, compositeRgba, getContrastRatio, resolveContrastColors } from './compositing';
+export {
+  BEBOP_WARM_REFERENCE_BACKDROP,
+  getHoverColor,
+  getInteractionColor,
+  getInteractionColorOverrides,
+  getInteractionColors,
+  getPressColor,
+  INTERACTION_COLOR_DELTAS,
+  INTERACTIVE_COLOR_TOKENS,
+  isInverseInteractionToken,
+} from './interaction';
