+++ packages/agentic/design/src/color-validation/index.ts
@@ -1,0 +1,12 @@
+export type {
+  ContrastColorDiagnostic,
+  ContrastColorToken,
+  ContrastPairing,
+  ContrastPairResult,
+  ContrastTokenReference,
+  ResolvedContrastPairResult,
+  UnresolvableContrastPairResult,
+} from './types';
+export { MODE_SURFACE, WCAG } from './constants';
+export { PAIRINGS } from './pairings';
+export { getContrastFailures, getUnresolvableContrastPairs, validateContrastPairs } from './validation';
