+++ packages/agentic/design/src/color-validation/testing/x3InteractionConformance.ts
@@ -1,4 +1,4 @@
-import type { InteractiveColorToken } from '../runtime-colors';
+import type { InteractiveColorToken } from '../../color-lib';
 
 export type InteractionConformanceVector = readonly [
   lightRest: string,
@@ -73,6 +73,69 @@ export const x3InteractionConformance = {
   foregroundWarningOnloud: ['#ffffff', '#ffffff', '#ffffff', '#000000db', '#000000db', '#000000db'],
   foregroundSuccessPrimary: ['#017048', '#006740', '#005f38', '#5dce97', '#67d8a0', '#72e2aa'],
   foregroundSuccessOnloud: ['#ffffff', '#ffffff', '#ffffff', '#000000db', '#000000db', '#000000db'],
+} as const satisfies Record<InteractiveColorToken, InteractionConformanceVector>;
+
+/** Base and generated interaction values for the pinned Bebop Warm theme. */
+export const x3BebopWarmInteractionConformance = {
+  surfaceNeutralFarther: ['#f0eae5', '#e8e0d9', '#dfd6ce', '#1e1916', '#342d29', '#4b423d'],
+  surfaceNeutralFar: ['#faf4f0', '#f2eae4', '#e9dfd9', '#231e1a', '#38312b', '#4d453e'],
+  surfaceNeutralNear: ['#f6f1ec', '#ede7e0', '#e5ddd4', '#1c1714', '#322b27', '#4b413c'],
+  surfaceNeutralNearer: ['#ffffff', '#f6f5f3', '#edebe7', '#2e2823', '#403832', '#534941'],
+  surfaceNeutralTranslucent: ['#ffffff80', '#f7f6f48a', '#f0eeea94', '#00000080', '#02010188', '#0e0b0c8f'],
+  backgroundNeutralHeavy: ['#28231e', '#3b352e', '#50473f', '#ffffff', '#f6f5f3', '#edebe7'],
+  backgroundNeutralLoud: ['#1e0d009a', '#2e1a0a9a', '#3f28169a', '#ffeee39a', '#f3dfd29a', '#e6d0c09a'],
+  backgroundNeutralSoft: ['#4d200028', '#61330f32', '#6a3b193c', '#ffe3cb40', '#fde1c548', '#ffdec44f'],
+  backgroundNeutralSubtle: ['#6c31001a', '#7f471c24', '#8b53272e', '#fff4ef14', '#ffe8e81c', '#ffdfdf23'],
+  backgroundNeutralTransparent: ['#28231e00', '#231d170a', '#1e170f14', '#ffffff00', '#fbfaf808', '#fefcf80f'],
+  backgroundNeutralTranslucent: ['#e9e9e980', '#dcdcd68a', '#d2cfc894', '#3c3c3c80', '#52524e88', '#67665e8f'],
+  backgroundBrandHeavy: ['#28231e', '#3b352e', '#50473f', '#ffffff', '#f6f5f3', '#edebe7'],
+  backgroundBrandLoud: ['#1e0d009a', '#2e1a0a9a', '#3f28169a', '#ffeee39a', '#f3dfd29a', '#e6d0c09a'],
+  backgroundBrandSoft: ['#4d200028', '#61330f32', '#6a3b193c', '#ffe3cb40', '#fde1c548', '#ffdec44f'],
+  backgroundBrandSubtle: ['#6c31001a', '#7f471c24', '#8b53272e', '#fff4ef14', '#ffe8e81c', '#ffdfdf23'],
+  backgroundBrandTransparent: ['#28231e00', '#231d170a', '#1e170f14', '#ffffff00', '#fbfaf808', '#fefcf80f'],
+  backgroundDangerLoud: ['#c02e56', '#cd355e', '#da3c65', '#e55e7a', '#dd5170', '#d54467'],
+  backgroundDangerSoft: ['#ffe3e6', '#f7d8dc', '#efced1', '#5a2932', '#68343d', '#773f49'],
+  backgroundDangerSubtle: ['#fff1f3', '#f7e6e9', '#efdcde', '#402529', '#523338', '#644247'],
+  backgroundWarningLoud: ['#bb4001', '#c84709', '#d54e11', '#e56738', '#dd5b28', '#d54f13'],
+  backgroundWarningSoft: ['#ffe5dc', '#f7dbd1', '#eed0c6', '#592c1c', '#673726', '#764331'],
+  backgroundWarningSubtle: ['#fff2ee', '#f7e7e3', '#efddd7', '#40261d', '#52352b', '#644439'],
+  backgroundSuccessLoud: ['#008455', '#088e5c', '#119864', '#05ad72', '#00a468', '#009c5e'],
+  backgroundSuccessSoft: ['#c9f8dd', '#bdefd3', '#b2e5c8', '#0c462d', '#185338', '#246043'],
+  backgroundSuccessSubtle: ['#e3fcee', '#d7f3e3', '#cce9d9', '#193426', '#274435', '#355544'],
+  strokeNeutralHeavy: ['#28231e', '#3b352e', '#50473f', '#e3dcd7', '#dbd2cb', '#d3c8c0'],
+  strokeNeutralLoud: ['#776d65', '#81756c', '#8c7e73', '#9a9089', '#92877e', '#8b7d74'],
+  strokeNeutralSoft: ['#250f0076', '#2a120280', '#2e16048a', '#ffe6d35c', '#ffe6cf64', '#ffe8d06b'],
+  strokeNeutralSubtle: ['#4d200028', '#61330f32', '#6a3b193c', '#ffe3cb40', '#fde1c548', '#ffdec44f'],
+  strokeNeutralTransparent: ['#28231e00', '#231d170a', '#1e170f14', '#ffffff00', '#fbfaf808', '#fefcf80f'],
+  strokeNeutralOnloud: ['#ffffffff', '#fffffd', '#fffffb', '#0b0600e1', '#000000e1', '#000000e1'],
+  strokeBrandLoud: ['#776d65', '#81756c', '#8c7e73', '#9a9089', '#92877e', '#8b7d74'],
+  strokeBrandSoft: ['#250f0076', '#2a120280', '#2e16048a', '#ffe6d35c', '#ffe6cf64', '#ffe8d06b'],
+  strokeBrandSubtle: ['#4d200028', '#61330f32', '#6a3b193c', '#ffe3cb40', '#fde1c548', '#ffdec44f'],
+  strokeBrandOnloud: ['#ffffffff', '#fffffd', '#fffffb', '#0b0600e1', '#000000e1', '#000000e1'],
+  strokeDangerLoud: ['#c02e56', '#cd355e', '#da3c65', '#e55e7a', '#dd5170', '#d54467'],
+  strokeDangerSoft: ['#e55e7a', '#dd5170', '#d54467', '#d14164', '#de486c', '#eb4f74'],
+  strokeDangerSubtle: ['#ffd0d6', '#f7c5cc', '#eebbc2', '#76293a', '#813041', '#8c3749'],
+  strokeDangerOnloud: ['#ffffff', '#fffffd', '#fffffb', '#242424', '#141312', '#060504'],
+  strokeWarningLoud: ['#cd4808', '#da4f11', '#e75618', '#e56738', '#dd5b28', '#d54f13'],
+  strokeWarningSoft: ['#fc9571', '#f48a65', '#ec7f58', '#cd4808', '#da4f11', '#e75618'],
+  strokeWarningSubtle: ['#ffd3c4', '#f7c9b9', '#eebeae', '#762e12', '#813518', '#8c3c1e'],
+  strokeWarningOnloud: ['#ffffff', '#fffffd', '#fffffb', '#242424', '#141312', '#060504'],
+  strokeSuccessLoud: ['#008455', '#088e5c', '#119864', '#05ad72', '#00a468', '#009c5e'],
+  strokeSuccessSoft: ['#5dce97', '#4cc58d', '#39bc82', '#008e5c', '#089863', '#11a26b'],
+  strokeSuccessSubtle: ['#b1efcc', '#a5e6c2', '#9adcb7', '#025636', '#0a5f3d', '#136844'],
+  strokeSuccessOnloud: ['#ffffff', '#fffffd', '#fffffb', '#242424', '#141312', '#060504'],
+  foregroundNeutralPrimary: ['#0b0600e1', '#060300eb', '#040100f5', '#fff7f2e3', '#fff8f2eb', '#fffaf2f2'],
+  foregroundNeutralSecondary: ['#190c00ab', '#0a0100b5', '#060000bf', '#fff2e8b5', '#fff7ebbd', '#fff9ecc4'],
+  foregroundNeutralTertiary: ['#1e0d009a', '#0c0100a4', '#080000ae', '#ffeee39a', '#fff6eaa2', '#fff8eaa9'],
+  foregroundNeutralOnloud: ['#ffffffff', '#fffffd', '#fffffb', '#0b0600e1', '#010000e1', '#000000e1'],
+  foregroundBrandPrimary: ['#0b0600e1', '#060300eb', '#040100f5', '#fff7f2e3', '#fff8f2eb', '#fffaf2f2'],
+  foregroundBrandOnloud: ['#ffffffff', '#fffffd', '#fffffb', '#0b0600e1', '#010000e1', '#000000e1'],
+  foregroundDangerPrimary: ['#a62147', '#9e0b3e', '#950035', '#fd8ea1', '#ff96aa', '#ff9db2'],
+  foregroundDangerOnloud: ['#ffffff', '#fffffd', '#fffffb', '#000000db', '#000000db', '#000000db'],
+  foregroundWarningPrimary: ['#a93901', '#a12c00', '#991c00', '#fc9571', '#ff9d77', '#ffa57d'],
+  foregroundWarningOnloud: ['#ffffff', '#fffffd', '#fffffb', '#000000db', '#000000db', '#000000db'],
+  foregroundSuccessPrimary: ['#017048', '#00683f', '#006036', '#5dce97', '#62d99f', '#67e4a7'],
+  foregroundSuccessOnloud: ['#ffffff', '#fffffd', '#fffffb', '#000000db', '#000000db', '#000000db'],
 } as const satisfies Record<InteractiveColorToken, InteractionConformanceVector>;
 
 /**
